Refactoring and restructuring code and tests
