+++ LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright (c) 2017 Kyle Swanson
+Copyright (c) 2019 Kyle Swanson
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
